Replaced byte[] with string in DTOs. Added JWT username claim. Added Username check from JWT to AlbumController.

+++ MusicPlayerBackend/Controllers/AlbumController.cs
@@ -32,7 +32,7 @@ namespace MusicPlayerBackend.Controllers
         public async Task<ActionResult<AlbumDTO>> AddAlbum(AlbumDTO album)
         {
             if (!_validator.IsAlbumDTOValid(album)) return BadRequest("Invalid album");
-            //Check here from JWT if the album user is same as JWT login user
+            if (this.User.Claims.First(i => i.Type == "Username").Value != album.UserName) return BadRequest("Invalid user for added album");
             if (await _albumRepository.CheckIfAlbumExists(album.Id)) return BadRequest("Album already exists");
 
             var albumResult = await _albumRepository.AddAsync(album);
@@ -46,7 +46,7 @@ namespace MusicPlayerBackend.Controllers
         public async Task<ActionResult<AlbumDTO>> UpdateAlbum(AlbumDTO album)
 		{
             if (!_validator.IsAlbumDTOValid(album)) return BadRequest("Invalid album");
-            //Check here from JWT if the album user is same as JWT login user
+            if (this.User.Claims.First(i => i.Type == "Username").Value != album.UserName) return BadRequest("Invalid user for added album");
             if (!await _albumRepository.CheckIfAlbumExists(album.Id)) return NotFound("Album you're trying to update does not exist");
 
             var albumResult = await _albumRepository.UpdateAsync(album);
@@ -60,7 +60,7 @@ namespace MusicPlayerBackend.Controllers
         public async Task<ActionResult<string>> DeleteAlbum(AlbumDTO album)
 		{
             if (!_validator.IsAlbumDTOValid(album)) return BadRequest("Invalid album");
-            //Check here from JWT if the album user is same as JWT login user
+            if (this.User.Claims.First(i => i.Type == "Username").Value != album.UserName) return BadRequest("Invalid user for added album");
             if (!await _albumRepository.CheckIfAlbumExists(album.Id)) return NotFound("Album you're trying to delete does not exist");
 
             var result = await _albumRepository.DeleteAsync(album.Id.ToString());
